Add support for calibration messages
Also fixed rust_anlyzer in the relay server's dev shell

+++ backend/relay/src/config.rs
@@ -7,15 +7,15 @@ pub const MQTT_PORT: u16 = 1883;
 pub const WS_PORT: u16 = 8443;
 
 // how many times per second that data is being sent from the server
-pub const FREQUENCY: u64 = 25;
+pub const FREQUENCY: usize = 25;
 
 // capacity of the channel which holds incoming messages for the mqtt client, as
 // well as the broadcast channel for communicating the the tasks handling
 // websocket connections.
-pub const CHANNEL_CAPACITY: usize = FREQUENCY as usize * 10;
+pub const CHANNEL_CAPACITY: usize = FREQUENCY * 10;
 
 // capacity of the cache of recent data that's sent to client when they first connect
-pub const CACHE_CAPACITY: usize = FREQUENCY as usize * 20;
+pub const CACHE_CAPACITY: usize = FREQUENCY * 20;
 
 // how many cm tall a wave needs to be to trigger an alert
 pub const ALERT_THRESHOLD: f32 = 8.0;
